Extracted batch dividing functionality into separate module
