+++ lib/worker/batcher.ex
@@ -24,6 +24,7 @@ defmodule BorsNG.Worker.Batcher do
 
   use GenServer
   alias BorsNG.Worker.Batcher
+  alias BorsNG.Worker.Batcher.Divider
   alias BorsNG.Database.Repo
   alias BorsNG.Database.Batch
   alias BorsNG.Database.BatchState
@@ -450,9 +451,13 @@ defmodule BorsNG.Worker.Batcher do
   end
 
   defp start_waiting_merged_batch(batch, patch_links, _base, :conflict) do
-    repo_conn = get_repo_conn(batch.project)
+    project = batch.project
+    repo_conn = get_repo_conn(project)
     patches = Enum.map(patch_links, &(&1.patch))
-    state = bisect_for_conflict_batch(patch_links, batch)
+    state = Divider.split_batch_with_conflicts(patch_links, batch)
+    if state == :retrying do
+      poll_after_delay(project)
+    end
 
     send_message(repo_conn, patches, {:conflict, state})
 
@@ -551,7 +556,10 @@ defmodule BorsNG.Worker.Batcher do
     |> LinkPatchBatch.from_batch()
     |> Repo.all()
     patches = Enum.map(patch_links, &(&1.patch))
-    state = bisect_for_failed_batch(patch_links, batch)
+    state = Divider.split_batch(patch_links, batch)
+    if state == :retrying do
+      poll_after_delay(project)
+    end
     send_message(repo_conn, patches, {state, erred})
   end
 
@@ -578,7 +586,10 @@ defmodule BorsNG.Worker.Batcher do
     |> LinkPatchBatch.from_batch()
     |> Repo.all()
     patches = Enum.map(patch_links, &(&1.patch))
-    state = bisect_for_failed_batch(patch_links, batch)
+    state = Divider.split_batch(patch_links, batch)
+    if state == :retrying do
+      poll_after_delay(project)
+    end
     project
     |> get_repo_conn()
     |> send_message(patches, {:timeout, state})
@@ -616,7 +627,7 @@ defmodule BorsNG.Worker.Batcher do
       uncanceled_patch_links = Enum.filter(
         patch_links,
         &(&1.patch_id != patch_id))
-      clone_batch(uncanceled_patch_links, project.id, batch.into_branch)
+      Divider.clone_batch(uncanceled_patch_links, project.id, batch.into_branch)
       canceled_patches = Enum.filter(
         patches,
         &(&1.id == patch_id))
@@ -643,82 +654,6 @@ defmodule BorsNG.Worker.Batcher do
     repo_conn = get_repo_conn(project)
     send_status(repo_conn, batch.id, [patch], :canceled)
     send_message(repo_conn, [patch], {:canceled, :failed})
-  end
-
-  defp bisect(patch_links, project_id, into) do
-    count = Enum.count(patch_links)
-
-    {lo, hi} = Enum.split(patch_links, div(count, 2))
-    clone_batch(lo, project_id, into)
-    clone_batch(hi, project_id, into)
-  end
-
-  defp isolate_unmergeable_patch_links(patch_links, repo_conn) do
-    patch_link_map =
-      patch_links
-      |> Enum.group_by (fn patch_link -> is_patch_mergeable(patch_link.patch, repo_conn) end)
-
-    {patch_link_map[true]||[], patch_link_map[false]||[]}
-  end
-
-  defp is_patch_mergeable(patch, repo_conn) do
-    pr = GitHub.get_pr!(repo_conn, patch.pr_xref)
-    pr.mergeable == true || pr.mergeable == nil
-  end
-
-  defp bisect_for_conflict_batch(patch_links, %Batch{project: project, into_branch: into}) do
-    repo_conn = get_repo_conn(project)
-
-    # if mergeable 0 and unmergeable = 1 -> fail no retry
-    #              0                   2+  create single batches for unmergeable patches
-    #              1                   0  impossible
-    #              1                   1+ create single batches for both
-    #              2+                  0  bisect for mergeable patches
-    #              2+                  1+ one batch for mergeable patches, create single batches for unmergeable patches
-    # Create batches for unmergeable patches first, so they will be picked up first and fail first.
-    case isolate_unmergeable_patch_links(patch_links, repo_conn) do
-      {[], [single_unmergeable]} ->
-        :failed
-
-      {[], multiple_unmergeable} ->
-        Enum.each multiple_unmergeable, fn patch_link ->
-          clone_batch([patch_link], project.id, into)
-        end
-        poll_after_delay(project)
-        :retrying
-
-      {[single_mergeable], multiple_unmergeable} ->
-        Enum.each multiple_unmergeable, fn patch_link ->
-          clone_batch([patch_link], project.id, into)
-        end
-        clone_batch([single_mergeable], project.id, into)
-        poll_after_delay(project)
-        :retrying
-
-      {multiple_mergeable, []} ->
-        bisect(multiple_mergeable, project.id, into)
-        poll_after_delay(project)
-        :retrying
-
-      {multiple_mergeable, multiple_unmergeable} ->
-        Enum.each multiple_unmergeable, fn patch_link ->
-          clone_batch([patch_link], project.id, into)
-        end
-        clone_batch(multiple_mergeable, project.id, into)
-        poll_after_delay(project)
-        :retrying
-    end
-  end
-
-  defp bisect_for_failed_batch(patch_links, %Batch{project: project, into_branch: into}) do
-    count = Enum.count(patch_links)
-    if count > 1 do
-      bisect(patch_links, project.id, into)
-      poll_after_delay(project)
-      :retrying
-    else
-      :failed
-    end
   end
 
   defp patch_preflight(repo_conn, patch) do
@@ -861,18 +796,6 @@ defmodule BorsNG.Worker.Batcher do
       review_required? ->
         :insufficient
     end
-  end
-
-  defp clone_batch(patch_links, project_id, into_branch) do
-    batch = Repo.insert!(Batch.new(project_id, into_branch))
-    patch_links
-    |> Enum.map(&%{
-      batch_id: batch.id,
-      patch_id: &1.patch_id,
-      reviewer: &1.reviewer})
-    |> Enum.map(&LinkPatchBatch.changeset(%LinkPatchBatch{}, &1))
-    |> Enum.each(&Repo.insert!/1)
-    batch
   end
 
   def get_new_batch(project_id, into_branch, priority) do
